requirements.txt
selenium
webdriver-manager
pytest
allure-pytest